+++ Runtime/Beam/Api/BeamPlayerClient/src/BeamPlayerClient/Model/GenerateSessionRequestResponse.cs
@@ -81,7 +81,7 @@ namespace BeamPlayerClient.Model
         /// <param name="address">address (required).</param>
         /// <param name="url">url (required).</param>
         [UnityEngine.Scripting.Preserve]
-        public GenerateSessionRequestResponse(StatusEnum status = default(StatusEnum), string id = default(string), DateTime createdAt = default(DateTime), DateTime? updatedAt = default(DateTime?), int chainId = default(int), string openfortId = default(string), string address = default(string), string url = default(string))
+        public GenerateSessionRequestResponse(StatusEnum status = default(StatusEnum), string id = default(string), DateTime createdAt = default(DateTime), DateTime? updatedAt = default(DateTime?), long chainId = default(long), string openfortId = default(string), string address = default(string), string url = default(string))
         {
             this.Status = status;
             // to ensure "id" is required (not null)
@@ -144,7 +144,7 @@ namespace BeamPlayerClient.Model
         /// </summary>
         [DataMember(Name = "chainId", IsRequired = true, EmitDefaultValue = true)]
         [UnityEngine.Scripting.Preserve]
-        public int ChainId { get; set; }
+        public long ChainId { get; set; }
 
         /// <summary>
         /// Gets or Sets OpenfortId
